Split audio workspace runtime hooks
