+++ tests/audio-workspace-architecture.test.ts
@@ -10,6 +10,8 @@ const controlsPath = join(audioDir, '_components/audio-workspace-controls.tsx');
 const generatedPickerPath = join(audioDir, '_components/audio-generated-video-picker.tsx');
 const helpersPath = join(audioDir, '_lib/audio-workspace-helpers.ts');
 const typesPath = join(audioDir, '_lib/audio-workspace-types.ts');
+const activeJobHookPath = join(audioDir, '_hooks/useAudioActiveJobPolling.ts');
+const generatedVideosHookPath = join(audioDir, '_hooks/useAudioGeneratedVideos.ts');
 
 const workspaceSource = readFileSync(workspacePath, 'utf8');
 
@@ -18,9 +20,13 @@ test('audio workspace delegates local controls, helpers, and contracts', () => {
   assert.ok(existsSync(generatedPickerPath), 'generated video picker should live in a route-local component module');
   assert.ok(existsSync(helpersPath), 'audio browser/API helpers should live in a route-local helper module');
   assert.ok(existsSync(typesPath), 'audio local contracts should live in a route-local type module');
+  assert.ok(existsSync(activeJobHookPath), 'active audio job polling should live in a route-local hook');
+  assert.ok(existsSync(generatedVideosHookPath), 'generated video loading should live in a route-local hook');
 
   assert.match(workspaceSource, /from '\.\/_components\/audio-workspace-controls'/);
   assert.match(workspaceSource, /from '\.\/_components\/audio-generated-video-picker'/);
+  assert.match(workspaceSource, /from '\.\/_hooks\/useAudioActiveJobPolling'/);
+  assert.match(workspaceSource, /from '\.\/_hooks\/useAudioGeneratedVideos'/);
   assert.match(workspaceSource, /from '\.\/_lib\/audio-workspace-helpers'/);
   assert.match(workspaceSource, /from '\.\/_lib\/audio-workspace-types'/);
 });
@@ -35,9 +41,11 @@ test('audio workspace does not regain extracted ownership', () => {
   assert.doesNotMatch(workspaceSource, /function AudioSelectControl\(/, 'audio controls belong in _components/audio-workspace-controls.tsx');
   assert.doesNotMatch(workspaceSource, /generated-source-skeleton/, 'generated video picker UI belongs in _components/audio-generated-video-picker.tsx');
   assert.doesNotMatch(workspaceSource, /copy\.picker\.audioBadge/, 'generated video picker card UI belongs in _components/audio-generated-video-picker.tsx');
+  assert.doesNotMatch(workspaceSource, /getJobStatus/, 'active job polling belongs in useAudioActiveJobPolling');
+  assert.doesNotMatch(workspaceSource, /surface=video&limit=60/, 'generated video loading belongs in useAudioGeneratedVideos');
 
   const lineCount = workspaceSource.split('\n').length;
-  assert.ok(lineCount <= 1160, `AudioWorkspace should stay below 1160 lines after generated picker extraction, got ${lineCount}`);
+  assert.ok(lineCount <= 1075, `AudioWorkspace should stay below 1075 lines after runtime hook extraction, got ${lineCount}`);
 });
 
 test('audio helper modules expose the expected workspace contract', () => {
@@ -45,6 +53,8 @@ test('audio helper modules expose the expected workspace contract', () => {
   const generatedPickerSource = readFileSync(generatedPickerPath, 'utf8');
   const helpersSource = readFileSync(helpersPath, 'utf8');
   const typesSource = readFileSync(typesPath, 'utf8');
+  const activeJobHookSource = readFileSync(activeJobHookPath, 'utf8');
+  const generatedVideosHookSource = readFileSync(generatedVideosHookPath, 'utf8');
 
   for (const exportName of [
     'AudioModePicker',
@@ -59,6 +69,10 @@ test('audio helper modules expose the expected workspace contract', () => {
     /export function AudioGeneratedVideoPickerModal/,
     'AudioGeneratedVideoPickerModal should be exported'
   );
+  assert.match(activeJobHookSource, /export function useAudioActiveJobPolling/, 'active job polling hook should be exported');
+  assert.match(activeJobHookSource, /getJobStatus/, 'active job polling hook should poll job status');
+  assert.match(generatedVideosHookSource, /export function useAudioGeneratedVideos/, 'generated videos hook should be exported');
+  assert.match(generatedVideosHookSource, /surface=video&limit=60/, 'generated videos hook should load video jobs');
 
   for (const exportName of [
     'DEFAULT_PACK',
